Completed change username nominal path, and refactored fake classes to test utils
